requirements.txt
livekit-agents[silero,turn-detector]~=1.3
livekit-plugins-noise-cancellation~=0.2
openai
python-dotenv
chromadb
# uv run agent.py download-files